requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
gunicorn==22.0.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
qdrant-client==1.9.1
fastembed==0.3.1
groq==0.11.0
httpx==0.27.0
esprima==4.0.1
numpy==1.26.4
slowapi==0.1.9
psycopg2-binary==2.9.9